feat!: use js-car blockstore interface (#5)
BREAKING CHANGE: The blockstore interface used by this library has been updated. If passing a blockstore to the constructor it should return `undefined` if the block is not present in the blockstore. If it is present, it should return a `{ bytes: Uint8Array, cid: CID }` object (a "Block"). Secondly the `getBlock` method now returns a `Block` instead of a `Uint8Array`.

resolves https://github.com/web3-storage/dagula/issues/3