feat!: support for multiple hash types (#9)
This PR adds support for multiple hash types (like `blake2b`, as used by
Filecoin).

To do this we make use of the "prefix" sent in bitswap wantlist
responses. The prefix contains just the metadata of the CID. Instead of
_assuming_ the multihash hasher is `sha256`, we now inspect the prefix
and hash using the specified hasher (if it exists). By default, only
`sha256` is supported (for smaller bundle), but in the CLI we also add
`blake2b`.

Secondarily, this PR fixes a memory leak with many listeners being added
to a passed abort controller which were never removed.

Thirdly, this PR moves the static `fromNetwork` factory function from
the `Dagula` class into `p2p.js` (as an export). This allows users who
are not using bitswap/libp2p (like freeway) to not end up with a massive
bundle.