feat!: add block streaming interface (#44)
This PR adds a `streamBlock(cid, [options])` method to Dagula that
returns a `ReadableStream` of block bytes. You can optionally pass a
byte range to extract a subset of bytes.

Also `statBlock(cid, [options])` that retrieves block info i.e. total
size in bytes.

The idea is to allow dagula to serve byte ranges from big blocks without
having to load the whole thing into memory.

BREAKING CHANGE: The `Blockstore` interface now includes a `stream`
method that returns a `ReadableStream` and a `stat` method that returns
info (byte size).

The format for the `entityBytes` option of `getPath` has changed. It is
now an array of 1 or 2 numbers. See `Range` type for more details.

refs
https://github.com/web3-storage/freeway/pull/100#discussion_r1591594799